requirements.txt
streamlit==1.32.2
plotly==5.22.0
pandas==2.0.3
folium==0.16.0
haversine==2.8.0
streamlit-folium==0.18.0
Pillow==10.0.0